Fixed issue #1.
