+++ last_date.txt
@@ -1,0 +1,1 @@
+Wed Dec 31 1969 18:00:00 GMT-0600 (CST)
