+++ lib/gmail/README.md
@@ -1,0 +1,8 @@
+# gmail
+The best app ever..
+
+### Useage:
+
+```
+// Sample code here.
+```
