+++ lib/notify/README.md
@@ -1,0 +1,8 @@
+# notify
+The best app ever..
+
+### Useage:
+
+```
+// Sample code here.
+```
